shared_libs/otel_lib/pyproject.toml
[build-system]
requires = ["setuptools>=82.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "otel-lib"
version = "0.1.0"
description = "OpenTelemetry distributed tracing library for Integration Hub"
authors = [{name = "DHCW"}]
requires-python = ">=3.13"
readme = "README.md"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "azure-monitor-opentelemetry==1.6.13",
    "opentelemetry-api>=1.29.0",
    "opentelemetry-sdk>=1.29.0",
    "urllib3>=2.6.3",
    "setuptools>=82.0.0",
]

[dependency-groups]
dev = [
  "ruff==0.15.0",
  "bandit==1.9.2",
  "mypy==1.18.2",
]

[tool.mypy]
python_version = "3.13"
disallow_untyped_defs = true

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
  "E",
  "F",
  "W",
  "A",
  "PLC",
  "PLE",
  "PLW",
  "I",
]

[tool.bandit.assert_used]
skips = ["*/test_*.py", "*/*_test.py"]

[tool.setuptools.packages.find]
where = ["."]
include = ["otel_lib*"]
